Crate rustc_codegen_ssa

This crate contains codegen code that is used by all codegen backends (LLVM and others). The backend-agnostic functions of this crate use functions defined in various traits that have to be implemented by each backend.

Modules

back

base

codegen_attrs

common

debuginfo

errors

Errors emitted by codegen_ssa

glue

meth

mir

mono_item

target_features

traits

Interface of a Rust codegen backend

Structs

CachedModuleCodegen

CodegenResults

CompiledModule

CrateInfo

Misc info we load from metadata to persist beyond the tcx.

MemFlags

ModuleCodegen

NativeLib

Enums

CodegenErrors

ModuleKind

Constants

RLINK_MAGIC 🔒

RLINK_VERSION 🔒

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Functions

looks_like_rust_object_file

Checks if the given filename ends with the .rcgu.o extension that rustc uses for the object files it generates.

provide

provide_extern