rustc_codegen_ssa
Module back

Modules

apple

archive

command 🔒

A thin wrapper around Command in the standard library which allows us to read the arguments that are built up.

link

linker 🔒

lto

metadata

Reading of the rustc metadata for rlibs and dylibs

rpath 🔒

symbol_export

write

Functions

versioned_llvm_target

The target triple depends on the deployment target, and is required to enable features such as cross-language LTO, and for picking the right Mach-O commands.