rustc_codegen_ssa::back
Module apple

Functions

add_version_to_llvm_target 🔒

deployment_target

Get the deployment target based on the standard environment variables, or fall back to the minimum version supported by rustc.

deployment_target_env_var

Name of the environment variable used to fetch the deployment target on the given OS.

macho_platform 🔒

minimum_deployment_target 🔒

The deployment target for the given target.

os_minimum_deployment_target 🔒

Minimum operating system versions currently supported by rustc.

parse_version 🔒

Parse an OS version triple (SDK version or deployment target).

pretty_version

Type Aliases

OSVersion 🔒

Deployment target or SDK version.