Function rustc_codegen_ssa::back::link::add_local_native_libraries

fn add_local_native_libraries(
    cmd: &mut dyn Linker,
    sess: &Session,
    codegen_results: &CodegenResults
)

Native library linking

User-supplied library search paths (-L on the command line). These are the same paths used to find Rust crates, so some of them may have been added already by the previous crate linking code. This only allows them to be found at compile time so it is still entirely up to outside forces to make sure that library can be found at runtime.

Also note that the native libraries linked here are only the ones located in the current crate. Upstream crates with native library dependencies may have their native library pulled in above.