Function rustc_codegen_ssa::back::link::linker_with_args

fn linker_with_args<'a>(
    path: &Path,
    flavor: LinkerFlavor,
    sess: &'a Session,
    archive_builder_builder: &dyn ArchiveBuilderBuilder,
    crate_type: CrateType,
    tmpdir: &Path,
    out_filename: &Path,
    codegen_results: &CodegenResults
) -> Result<Command, ErrorGuaranteed>

Produce the linker command line containing linker path and arguments.

When comments in the function say “order-(in)dependent” they mean order-dependence between options and libraries/object files. For example --whole-archive (order-dependent) applies to specific libraries passed after it, and -o (output file, order-independent) applies to the linking process as a whole. Order-independent options may still override each other in order-dependent fashion, e.g --foo=yes --foo=no may be equivalent to --foo=no.