Module rustc_codegen_ssa::back::linker

Structs

BpfLinker

EmLinker

GccLinker

L4Bender

Linker shepherd script for L4Re (Fiasco)

MsvcLinker

PtxLinker

Much simplified and explicit CLI for the NVPTX linker. The linker operates with bitcode and uses LLVM backend to generate a PTX assembly.

WasmLd

Traits

Linker

Linker abstraction used by back::link to build up the command to invoke a linker.

Functions

disable_localization

Disables non-English messages from localized linkers. Such messages may cause issues with text encoding on Windows (#35785) and prevent inspection of linker output in case of errors, which we occasionally do. This should be acceptable because other messages from rustc are in English anyway, and may also be desirable to improve searchability of the linker diagnostics.

exported_symbols 🔒

for_each_exported_symbols_include_dep 🔒

get_linker

linked_symbols 🔒