Struct rustc_codegen_ssa::back::write::ModuleConfig

pub struct ModuleConfig {

    pub passes: Vec<String>,
    pub opt_level: Option<OptLevel>,
    pub opt_size: Option<OptLevel>,
    pub pgo_gen: SwitchWithOptPath,
    pub pgo_use: Option<PathBuf>,
    pub pgo_sample_use: Option<PathBuf>,
    pub debug_info_for_profiling: bool,
    pub instrument_coverage: bool,
    pub instrument_gcov: bool,
    pub sanitizer: SanitizerSet,
    pub sanitizer_recover: SanitizerSet,
    pub sanitizer_memory_track_origins: usize,
    pub emit_pre_lto_bc: bool,
    pub emit_no_opt_bc: bool,
    pub emit_bc: bool,
    pub emit_ir: bool,
    pub emit_asm: bool,
    pub emit_obj: EmitObj,
    pub emit_thin_lto: bool,
    pub bc_cmdline: String,
    pub verify_llvm_ir: bool,
    pub no_prepopulate_passes: bool,
    pub no_builtins: bool,
    pub time_module: bool,
    pub vectorize_loop: bool,
    pub vectorize_slp: bool,
    pub merge_functions: bool,
    pub inline_threshold: Option<u32>,
    pub new_llvm_pass_manager: Option<bool>,
    pub emit_lifetime_markers: bool,
    pub llvm_plugins: Vec<String>,

}

Module-specific configuration for optimize_and_codegen.

Fields
passes: Vec<String>

Names of additional optimization passes to run.

opt_level: Option<OptLevel>

Some(level) to optimize at a certain level, or None to run absolutely no optimizations (used for the metadata module).

opt_size: Option<OptLevel>

Some(level) to optimize binary size, or None to not affect program size.

pgo_gen: SwitchWithOptPathpgo_use: Option<PathBuf>pgo_sample_use: Option<PathBuf>debug_info_for_profiling: boolinstrument_coverage: boolinstrument_gcov: boolsanitizer: SanitizerSetsanitizer_recover: SanitizerSetsanitizer_memory_track_origins: usizeemit_pre_lto_bc: boolemit_no_opt_bc: boolemit_bc: boolemit_ir: boolemit_asm: boolemit_obj: EmitObjemit_thin_lto: boolbc_cmdline: Stringverify_llvm_ir: boolno_prepopulate_passes: boolno_builtins: booltime_module: boolvectorize_loop: boolvectorize_slp: boolmerge_functions: boolinline_threshold: Option<u32>new_llvm_pass_manager: Option<bool>emit_lifetime_markers: boolllvm_plugins: Vec<String>
Implementations

impl ModuleConfig

fn new(
kind: ModuleKind,
sess: &Session,
no_builtins: bool,
is_compiler_builtins: bool
) -> ModuleConfig

pub fn bitcode_needed(&self) -> bool