Module rustc_codegen_ssa::base

Functions

bin_op_to_fcmp_predicate

bin_op_to_icmp_predicate

cast_shift_expr_rhs

cast_to_dyn_star

Coerces src to dst_ty which is guaranteed to be a dyn* type.

codegen_crate

codegen_instance

coerce_unsized_into

Coerces src, which is a reference to a value of type src_ty, to a value of type dst_ty, and stores the result in dst.

collect_debugger_visualizers_transitive

This function returns all of the debugger visualizers specified for the current crate as well as all upstream crates transitively that match the visualizer_type specified.

compare_simd_types

determine_cgu_reuse 🔒

get_argc_argv 🔒

Obtain the argc and argv values to pass to the rust start function.

maybe_create_entry_wrapper

Creates the main function which will initialize the rust runtime and call users main function.

memcpy_ty

provide

unsize_ptr

Coerces src to dst_ty. src_ty must be a pointer.

unsized_info

Retrieves the information we are losing (making dynamic) in an unsizing adjustment.

vtable_ptr_ty 🔒

wants_msvc_seh

Returns true if this session’s target will use SEH-based unwinding.