Function rustc_codegen_ssa::base::unsized_info

pub fn unsized_info<'a, 'tcx, Bx: BuilderMethods<'a, 'tcx>>(
    bx: &mut Bx,
    source: Ty<'tcx>,
    target: Ty<'tcx>,
    old_info: Option<Bx::Value>
) -> Bx::Value

Retrieves the information we are losing (making dynamic) in an unsizing adjustment.

The old_info argument is a bit odd. It is intended for use in an upcast, where the new vtable for an object will be derived from the old one.