rustc_codegen_ssa

Module meth

Structs

VirtualIndex 🔒

Functions

dyn_trait_in_self 🔒

This takes a valid self receiver type and extracts the principal trait ref of the type. Return None if there is no principal trait.

get_vtable 🔒

Creates a dynamic vtable for the given type and vtable origin. This is used only for objects.

load_vtable 🔒

Call this function whenever you need to load a vtable.