Module rustc_codegen_ssa::mir

Modules

analyze 🔒

An analysis to determine which locals require allocas and which do not.

block 🔒

constant

coverageinfo

debuginfo

intrinsic 🔒

operand

place

rvalue 🔒

statement 🔒

Structs

FunctionCx

Master context for codegenning from MIR.

Enums

LocalRef 🔒

Functions

arg_local_refs 🔒

Produces, for each argument, a Value pointing at the argument’s value. As arguments are places, these are always indirect.

codegen_mir