Module rustc_codegen_ssa::mir::analyze

An analysis to determine which locals require allocas and which do not.

Structs

LocalAnalyzer 🔒

Enums

CleanupKind

LocalKind 🔒

Functions

cleanup_kinds

MSVC requires unwinding code to be split to a tree of funclets, where each funclet can only branch to itself or to its parent. Luckily, the code we generates matches this pattern. Recover that structure in an analyze pass.

non_ssa_locals