Module rustc_codegen_ssa::target_features

Constants

AARCH64_ALLOWED_FEATURES 🔒

AARCH64_TIED_FEATURES 🔒

ARM_ALLOWED_FEATURES 🔒

BPF_ALLOWED_FEATURES 🔒

HEXAGON_ALLOWED_FEATURES 🔒

MIPS_ALLOWED_FEATURES 🔒

POWERPC_ALLOWED_FEATURES 🔒

RISCV_ALLOWED_FEATURES 🔒

RUSTC_SPECIFIC_FEATURES

Features that control behaviour of rustc, rather than the codegen.

WASM_ALLOWED_FEATURES 🔒

X86_ALLOWED_FEATURES 🔒

Functions

all_known_features

When rustdoc is running, provide a list of all known features so that all their respective primitives may be documented.

asm_target_features 🔒

Computes the set of target features used in a function for the purposes of inline assembly.

check_target_feature_trait_unsafe

Checks the function annotated with #[target_feature] is not a safe trait method implementation, reporting an error if it is.

from_target_feature

provide 🔒

supported_target_features

tied_target_features