Module rustc_codegen_ssa::back::linker

Macros

generate_arg_methods 🔒

Structs

AixLinker

Linker for AIX.

BpfLinker

EmLinker

GccLinker

L4Bender

Linker shepherd script for L4Re (Fiasco)

LlbcLinker

The self-contained LLVM bitcode linker

MsvcLinker

PtxLinker

Much simplified and explicit CLI for the NVPTX linker. The linker operates with bitcode and uses LLVM backend to generate a PTX assembly.

WasmLd

Traits

Linker

Linker abstraction used by back::link to build up the command to invoke a linker.

Functions

cc_args 🔒

Arguments for the cc wrapper specifically. Check that it’s indeed a cc wrapper and pass verbatim.

disable_localization

Disables non-English messages from localized linkers. Such messages may cause issues with text encoding on Windows (#35785) and prevent inspection of linker output in case of errors, which we occasionally do. This should be acceptable because other messages from rustc are in English anyway, and may also be desirable to improve searchability of the linker diagnostics.

exported_symbols 🔒

exported_symbols_for_non_proc_macro 🔒

exported_symbols_for_proc_macro_crate 🔒

for_each_exported_symbols_include_dep 🔒

get_linker

The third parameter is for env vars, used on windows to set up the path for MSVC to find its DLLs, and gcc to find its bundled toolchain

link_args 🔒

Arguments for the underlying linker. Add options to pass them through cc wrapper if Linker is a cc wrapper.

link_or_cc_args 🔒

Arguments supported by both underlying linker and cc wrapper, pass verbatim.

linked_symbols 🔒

verbatim_args 🔒

Just pass the arguments to the linker as is. It is assumed that they are correctly prepared in advance.