Struct rustc_codegen_ssa::back::linker::GccLinker

pub struct GccLinker<'a> {
    cmd: Command,
    sess: &'a Session,
    target_cpu: &'a str,
    hinted_static: bool,
    is_ld: bool,
}

Fields
cmd: Commandsess: &'a Sessiontarget_cpu: &'a strhinted_static: boolis_ld: bool
Implementations

impl<'a> GccLinker<'a>

fn linker_arg(&mut self, arg: impl AsRef<OsStr>) -> &mut Self

Passes an argument directly to the linker.

When the linker is not ld-like such as when using a compiler as a linker, the argument is prepended by -Wl,.

fn linker_args(&mut self, args: &[impl AsRef<OsStr>]) -> &mut Self

Passes a series of arguments directly to the linker.

When the linker is ld-like, the arguments are simply appended to the command. When the linker is not ld-like such as when using a compiler as a linker, the arguments are joined by commas to form an argument that is then prepended with -Wl. In this situation, only a single argument is appended to the command to ensure that the order of the arguments is preserved by the compiler.

fn takes_hints(&self) -> bool

fn hint_static(&mut self)

fn hint_dynamic(&mut self)

fn push_linker_plugin_lto_args(&mut self, plugin_path: Option<&OsStr>)

fn build_dylib(&mut self, out_filename: &Path)