rustc_codegen_ssa::back
Module symbol_export

Functions

crate_export_threshold ğŸ”’

crates_export_threshold

exported_symbols_provider_local ğŸ”’

exporting_symbol_name_for_instance_in_crate ğŸ”’

is_reachable_non_generic_provider_extern ğŸ”’

is_reachable_non_generic_provider_local ğŸ”’

is_unreachable_local_definition_provider ğŸ”’

linking_symbol_name_for_instance_in_crate ğŸ”’

This is the symbol name of the given instance as seen by the linker.

maybe_emutls_symbol_name ğŸ”’

provide ğŸ”’

reachable_non_generics_provider ğŸ”’

symbol_export_level ğŸ”’

symbol_name_for_instance_in_crate ğŸ”’

This is the symbol name of the given instance instantiated in a specific crate.

threshold ğŸ”’

upstream_async_drop_glue_for_provider ğŸ”’

upstream_drop_glue_for_provider ğŸ”’

upstream_monomorphizations_for_provider ğŸ”’

upstream_monomorphizations_provider ğŸ”’

wasm_import_module_map ğŸ”’