Module rustc_codegen_ssa::back::symbol_export

Functions

crate_export_threshold 🔒

crates_export_threshold

exported_symbols_provider_local 🔒

exporting_symbol_name_for_instance_in_crate

is_reachable_non_generic_provider_extern 🔒

is_reachable_non_generic_provider_local 🔒

is_unreachable_local_definition_provider 🔒

linking_symbol_name_for_instance_in_crate

This is the symbol name of the given instance as seen by the linker.

maybe_emutls_symbol_name 🔒

provide

reachable_non_generics_provider 🔒

symbol_export_level 🔒

symbol_name_for_instance_in_crate

This is the symbol name of the given instance instantiated in a specific crate.

threshold

upstream_async_drop_glue_for_provider 🔒

upstream_drop_glue_for_provider 🔒

upstream_monomorphizations_for_provider 🔒

upstream_monomorphizations_provider 🔒

wasm_import_module_map 🔒