rustc_codegen_ssa::back

Module symbol_export

Functions

calling_convention_for_symbol ğŸ”’

crate_export_threshold ğŸ”’

crates_export_threshold

exported_generic_symbols_provider_local ğŸ”’

exported_non_generic_symbols_provider_local ğŸ”’

exporting_symbol_name_for_instance_in_crate ğŸ”’

extend_exported_symbols ğŸ”’

On amdhsa, gpu-kernel functions have an associated metadata object with a .kd suffix. Add it to the symbols list for all kernel functions, so that it is exported in the linked object.

is_reachable_non_generic_provider_extern ğŸ”’

is_reachable_non_generic_provider_local ğŸ”’

is_unreachable_local_definition_provider ğŸ”’

linking_symbol_name_for_instance_in_crate ğŸ”’

This is the symbol name of the given instance as seen by the linker.

maybe_emutls_symbol_name ğŸ”’

provide ğŸ”’

reachable_non_generics_provider ğŸ”’

symbol_export_level ğŸ”’

symbol_name_for_instance_in_crate ğŸ”’

This is the symbol name of the given instance instantiated in a specific crate.

threshold ğŸ”’

upstream_async_drop_glue_for_provider ğŸ”’

upstream_drop_glue_for_provider ğŸ”’

upstream_monomorphizations_for_provider ğŸ”’

upstream_monomorphizations_provider ğŸ”’

wasm_import_module_map ğŸ”’