rustc_codegen_ssa::back
Module write

Structs

CguMessage

A message sent from the coordinator thread to the main thread telling it to process another codegen unit.

CodegenContext

Additional resources used by optimize_and_codegen (not module specific)

CompiledModules ğŸ”’

Coordinator

Diagnostic ğŸ”’

ModuleConfig

Module-specific configuration for optimize_and_codegen.

OngoingCodegen

SharedEmitter

SharedEmitterMain

Subdiagnostic ğŸ”’

TargetMachineFactoryConfig

Configuration passed to the function returned by the target_machine_factory.

WorkerFatalError ğŸ”’

FatalError is explicitly not Send.

Enums

BitcodeSection

What kind of llvm bitcode section to embed in an object file.

ComputedLtoType ğŸ”’

Actual LTO type we end up choosing based on multiple factors.

EmitObj

What kind of object file to emit.

FatLtoInput

MainThreadState ğŸ”’

Message ğŸ”’

Messages sent to the coordinator.

SharedEmitterMessage ğŸ”’

WorkItem ğŸ”’

WorkItemResult ğŸ”’

A result produced by the backend.

Constants

PRE_LTO_BC_EXT ğŸ”’

Functions

compute_per_cgu_lto_type ğŸ”’

copy_all_cgu_workproducts_to_incr_comp_cache_dir ğŸ”’

execute_copy_from_cache_work_item ğŸ”’

execute_lto_work_item ğŸ”’

execute_optimize_work_item ğŸ”’

finish_intra_module_work ğŸ”’

generate_lto_work ğŸ”’

msvc_imps_needed ğŸ”’

need_bitcode_in_object ğŸ”’

need_pre_lto_bitcode_for_incr_comp ğŸ”’

pre_lto_bitcode_filename ğŸ”’

produce_final_output_artifacts ğŸ”’

spawn_work ğŸ”’

start_async_codegen ğŸ”’

start_executing_work ğŸ”’

submit_codegened_module_to_llvm ğŸ”’

submit_post_lto_module_to_llvm ğŸ”’

submit_pre_lto_module_to_llvm ğŸ”’

Type Aliases

ExportedSymbols ğŸ”’

TargetMachineFactoryFn