Module rustc_codegen_ssa::back::write

Structs

CguMessage

A message sent from the coordinator thread to the main thread telling it to process another codegen unit.

CodegenContext

Additional resources used by optimize_and_codegen (not module specific)

CompiledModules

Coordinator

Diagnostic 🔒

ModuleConfig

Module-specific configuration for optimize_and_codegen.

OngoingCodegen

SharedEmitter

SharedEmitterMain

Subdiagnostic

TargetMachineFactoryConfig

Configuration passed to the function returned by the target_machine_factory.

WorkerFatalError

FatalError is explicitly not Send.

Enums

BitcodeSection

What kind of llvm bitcode section to embed in an object file.

ComputedLtoType

Actual LTO type we end up choosing based on multiple factors.

EmitObj

What kind of object file to emit.

FatLtoInput

MainThreadState 🔒

Message 🔒

Messages sent to the coordinator.

SharedEmitterMessage 🔒

WorkItem 🔒

WorkItemResult 🔒

A result produced by the backend.

Constants

PRE_LTO_BC_EXT 🔒

Functions

compute_per_cgu_lto_type

copy_all_cgu_workproducts_to_incr_comp_cache_dir 🔒

execute_copy_from_cache_work_item 🔒

execute_lto_work_item 🔒

execute_optimize_work_item 🔒

finish_intra_module_work 🔒

generate_lto_work 🔒

msvc_imps_needed 🔒

need_bitcode_in_object 🔒

need_pre_lto_bitcode_for_incr_comp 🔒

pre_lto_bitcode_filename 🔒

produce_final_output_artifacts 🔒

spawn_work 🔒

start_async_codegen

start_executing_work 🔒

submit_codegened_module_to_llvm

submit_post_lto_module_to_llvm

submit_pre_lto_module_to_llvm

Type Aliases

ExportedSymbols

TargetMachineFactoryFn