rustc_codegen_ssa
Module codegen_attrs

Structs

MixedExportNameAndNoMangleState ğŸ”’

Functions

check_link_name_xor_ordinal ğŸ”’

check_link_ordinal ğŸ”’

check_tied_features

Given a map from target_features to whether they are enabled or disabled, ensure only valid combinations are allowed.

codegen_fn_attrs ğŸ”’

linkage_by_name ğŸ”’

provide ğŸ”’

should_inherit_track_caller ğŸ”’

Checks if the provided DefId is a method in a trait impl for a trait which has track_caller applied to the method prototype.