rustc_codegen_ssa
Module debuginfo

Modules

type_names

Type Names for Debug Info.

Functions

tag_base_type

Extract the type with which we want to describe the tag of the given enum or coroutine.

tag_base_type_opt 🔒

wants_c_like_enum_debuginfo

Returns true if we want to generate a DW_TAG_enumeration_type description for this instead of a DW_TAG_struct_type with DW_TAG_variant_part.