Module rustc_codegen_ssa::debuginfo::type_names

Type Names for Debug Info.

Enums

VTableNameKind

Functions

compute_debuginfo_type_name

Compute the name of the type as it should be stored in debuginfo. Does not do any caching, i.e., calling the function twice with the same type will also do the work twice. The qualified parameter only affects the first level of the type name, further levels (i.e., type parameters) are always fully qualified.

compute_debuginfo_vtable_name

Computes a name for the global variable storing a vtable (or the type of that global variable).

coroutine_kind_label 🔒

cpp_like_debuginfo

Check if we should generate C++ like names and debug information.

pop_arg_separator 🔒

pop_close_angle_bracket 🔒

push_arg_separator 🔒

push_close_angle_bracket 🔒

push_closure_or_coroutine_name 🔒

push_const_param 🔒

push_debuginfo_type_name 🔒

push_disambiguated_special_name 🔒

push_generic_params

push_generic_params_internal 🔒

push_item_name

push_unqualified_item_name 🔒