rustc_codegen_ssa
Module mir

Modules

analyze 🔒

An analysis to determine which locals require allocas and which do not.

block 🔒

constant 🔒

coverageinfo 🔒

debuginfo

intrinsic 🔒

locals 🔒

Locals are in a private module as updating LocalRef::Operand has to be careful wrt to subtyping. To deal with this we only allow updates by using FunctionCx::overwrite_local which handles it automatically.

naked_asm 🔒

operand

place

rvalue 🔒

statement 🔒

Structs

FunctionCx

Master context for codegenning from MIR.

Enums

CachedLlbb 🔒

LocalRef 🔒

Functions

arg_local_refs 🔒

Produces, for each argument, a Value pointing at the argument’s value. As arguments are places, these are always indirect.

codegen_mir

find_cold_blocks 🔒

Type Aliases

PerLocalVarDebugInfoIndexVec 🔒