Struct rustc_codegen_ssa::mir::block::TerminatorCodegenHelper

struct TerminatorCodegenHelper<'tcx> {
    bb: BasicBlock,
    terminator: &'tcx Terminator<'tcx>,
}

Used by FunctionCx::codegen_terminator for emitting common patterns e.g., creating a basic block, calling a function, etc.

Fields
bb: BasicBlockterminator: &'tcx Terminator<'tcx>
Implementations

impl<'a, 'tcx> TerminatorCodegenHelper<'tcx>

fn funclet<'b, Bx: BuilderMethods<'a, 'tcx>>( &self, fx: &'b mut FunctionCx<'a, 'tcx, Bx> ) -> Option<&'b Bx::Funclet>

Returns the appropriate Funclet for the current funclet, if on MSVC, either already previously cached, or newly created, by landing_pad_for.

fn llbb_with_cleanup<Bx: BuilderMethods<'a, 'tcx>>( &self, fx: &mut FunctionCx<'a, 'tcx, Bx>, target: BasicBlock ) -> Bx::BasicBlock

Get a basic block (creating it if necessary), possibly with cleanup stuff in it or next to it.

fn llbb_characteristics<Bx: BuilderMethods<'a, 'tcx>>( &self, fx: &mut FunctionCx<'a, 'tcx, Bx>, target: BasicBlock ) -> (bool, bool)

fn funclet_br<Bx: BuilderMethods<'a, 'tcx>>( &self, fx: &mut FunctionCx<'a, 'tcx, Bx>, bx: &mut Bx, target: BasicBlock, mergeable_succ: bool ) -> MergingSucc

fn do_call<Bx: BuilderMethods<'a, 'tcx>>( &self, fx: &mut FunctionCx<'a, 'tcx, Bx>, bx: &mut Bx, fn_abi: &'tcx FnAbi<'tcx, Ty<'tcx>>, fn_ptr: Bx::Value, llargs: &[Bx::Value], destination: Option<(ReturnDest<'tcx, Bx::Value>, BasicBlock)>, unwind: UnwindAction, copied_constant_arguments: &[PlaceRef<'tcx, <Bx as BackendTypes>::Value>], mergeable_succ: bool ) -> MergingSucc

Call fn_ptr of fn_abi with the arguments llargs, the optional return destination destination and the unwind action unwind.

fn do_inlineasm<Bx: BuilderMethods<'a, 'tcx>>( &self, fx: &mut FunctionCx<'a, 'tcx, Bx>, bx: &mut Bx, template: &[InlineAsmTemplatePiece], operands: &[InlineAsmOperandRef<'tcx, Bx>], options: InlineAsmOptions, line_spans: &[Span], destination: Option<BasicBlock>, unwind: UnwindAction, instance: Instance<'_>, mergeable_succ: bool ) -> MergingSucc

Generates inline assembly with optional destination and unwind.