Trait rustc_codegen_ssa::traits::BaseTypeMethods

pub trait BaseTypeMethods<'tcx>: Backend<'tcx> {

    fn type_i1(&self) -> Self::Type;
    fn type_i8(&self) -> Self::Type;
    fn type_i16(&self) -> Self::Type;
    fn type_i32(&self) -> Self::Type;
    fn type_i64(&self) -> Self::Type;
    fn type_i128(&self) -> Self::Type;
    fn type_isize(&self) -> Self::Type;
    fn type_f32(&self) -> Self::Type;
    fn type_f64(&self) -> Self::Type;
    fn type_func(&self, args: &[Self::Type], ret: Self::Type) -> Self::Type;
    fn type_struct(&self, els: &[Self::Type], packed: bool) -> Self::Type;
    fn type_kind(&self, ty: Self::Type) -> TypeKind;
    fn type_ptr_to(&self, ty: Self::Type) -> Self::Type;
    fn type_ptr_to_ext(
        &self,
        ty: Self::Type,
        address_space: AddressSpace
    ) -> Self::Type;
    fn element_type(&self, ty: Self::Type) -> Self::Type;
    fn vector_length(&self, ty: Self::Type) -> usize;
    fn float_width(&self, ty: Self::Type) -> usize;
    fn int_width(&self, ty: Self::Type) -> u64;
    fn val_ty(&self, v: Self::Value) -> Self::Type;

}

Required Methods

fn type_i1(&self) -> Self::Type

fn type_i8(&self) -> Self::Type

fn type_i16(&self) -> Self::Type

fn type_i32(&self) -> Self::Type

fn type_i64(&self) -> Self::Type

fn type_i128(&self) -> Self::Type

fn type_isize(&self) -> Self::Type

fn type_f32(&self) -> Self::Type

fn type_f64(&self) -> Self::Type

fn type_func(&self, args: &[Self::Type], ret: Self::Type) -> Self::Type

fn type_struct(&self, els: &[Self::Type], packed: bool) -> Self::Type

fn type_kind(&self, ty: Self::Type) -> TypeKind

fn type_ptr_to(&self, ty: Self::Type) -> Self::Type

fn type_ptr_to_ext(
&self,
ty: Self::Type,
address_space: AddressSpace
) -> Self::Type

fn element_type(&self, ty: Self::Type) -> Self::Type

fn vector_length(&self, ty: Self::Type) -> usize

Returns the number of elements in self if it is a LLVM vector type.

fn float_width(&self, ty: Self::Type) -> usize

fn int_width(&self, ty: Self::Type) -> u64

Retrieves the bit width of the integer type self.

fn val_ty(&self, v: Self::Value) -> Self::Type